pyproject.toml
[tool.poetry]
name = "agentchat"
version = "2.2.0"
description = "An intelligent agent communication platform based on LLM"
authors = ["tianmingguang <2593666979@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^1.0.3"
langchain-community = "^0.4.1"
pdf2docx = "^0.5.8"
loguru = "^0.7.3"
langchain-openai = "^1.0.2"
dashscope = "^1.25.0"
tavily-python = "^0.7.12"
mcp = "^1.20.0"
fastapi = "^0.121.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
python-multipart = "^0.0.20"
sqlmodel = "^0.0.27"
pytz = "^2025.2"
redis = "^7.0.1"
oss2 = "^2.19.1"
pypandoc = "^1.15"
elasticsearch = "^9.2.0"
pymilvus = "^2.6.3"
pymupdf4llm = "^0.1.8"
aiofiles = "^25.1.0"
chromadb = "^1.3.4"
requests-html = "^0.10.0"
selenium = "^4.38.0"
pymysql = "^1.1.2"
aiomysql = "^0.3.2"
google-search-results = "^2.4.2"
arxiv = "^2.3.0"
anthropic = "^0.72.0"
html2image = "^2.0.7"
lxml-html-clean = "^0.4.3"
pyfiglet = "^1.0.4"
minio = "^7.2.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
